requirements.txt
pandas
numpy
scipy
jupyter
matplotlib
seaborn
scikit-learn
mlflow
pyyaml
streamlit
pydantic
fastapi
uvicorn
